requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
Pillow>=10.0
reportlab>=4.0
weasyprint>=60.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
twilio>=8.0
django-environ>=0.11
openpyxl>=3.1
django-extensions>=3.2
